+++ Rdmp.Core/Curation/TableInfoImporter.cs
@@ -119,7 +119,7 @@ namespace Rdmp.Core.Curation
             tableName = querySyntaxHelper.EnsureWrapped(_importDatabaseName);
 
             if (_type == DatabaseType.MicrosoftSQLServer || _type == DatabaseType.PostgreSql)
-                tableName += "."+(_importFromSchema ?? querySyntaxHelper.GetDefaultSchemaIfAny()) +".";
+                tableName += "."+(querySyntaxHelper.EnsureWrapped(_importFromSchema ?? querySyntaxHelper.GetDefaultSchemaIfAny())) +".";
             else if (_type == DatabaseType.MySql || _type == DatabaseType.Oracle)
                 tableName += ".";
             else